% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologer.R
\name{open.data}
\alias{open.data}
\title{Open Biologer Data}
\usage{
\method{open}{data}(path = "biologer_data.csv", verbose = TRUE)
}
\arguments{
\item{path}{A character string specifying the path to the CSV file.
If NULL, defaults to the path used in `get_biologer_data()`.}

\item{verbose}{Logical, whether to display messages about data loading. Defaults to TRUE.}
}
\value{
A dataframe containing the Biologer data.
}
\description{
Opens the Biologer data CSV file as a dataframe.
}
