% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologer.R
\name{polygon.counts}
\alias{polygon.counts}
\title{Count Observations Per Polygon}
\usage{
polygon.counts(fill = "taxon.name", polygon_coords, polygon_names = NULL, df)
}
\arguments{
\item{fill}{Column name to group and count rows by. Default is "taxon.name".}

\item{polygon_coords}{A data frame or list of data frames defining polygon(s) with longitude and latitude columns.}

\item{polygon_names}{Optional vector of names for the polygons. If not provided, polygons will be numbered.}

\item{df}{A dataframe containing latitude and longitude columns, as well as the column specified in fill.}
}
\value{
A dataframe with counts for each polygon and a TOTAL row.
}
\description{
Counts the number of observations that fall within each provided polygon.
}
