% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologer.R
\name{subset.biologer}
\alias{subset.biologer}
\title{Subset Biologer Data}
\usage{
\method{subset}{biologer}(df = NULL, ...)
}
\arguments{
\item{df}{A dataframe to subset. Defaults to the result of `open.data()` if NULL.}

\item{...}{Named arguments specifying column names and the values to subset by.
For example: `sex = "value"`, `taxon.name = "value"`, or
`taxon.id = c("value1", "value2")`.}
}
\value{
A subsetted dataframe.
}
\description{
Subset a Biologer dataframe based on specified column-value conditions.
}
