% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologer.R
\name{zupanije.biologer}
\alias{zupanije.biologer}
\title{Count Observations Per Zupanija}
\usage{
zupanije.biologer(fill = "taxon.name", df, selected_zupanije = NULL)
}
\arguments{
\item{fill}{Column name to group and count rows by. Default is "taxon.name".}

\item{df}{A dataframe containing latitude and longitude columns.}

\item{selected_zupanije}{Optional vector of županija names to filter by. Default is NULL (all županije).}
}
\value{
A dataframe with counts for each Zupanija and a TOTAL row.
}
\description{
Counts the number of observations per Zupanija based on the coordinates of rows in the dataframe.
}
